DolphinWX: Break the netplay UI controls into separate classes.
