+++ Source/Core/DolphinWX/NetPlay/ChangeGameDialog.h
@@ -1,0 +1,22 @@
+// Copyright 2015 Dolphin Emulator Project
+// Licensed under GPLv2
+// Refer to the license.txt file included.
+
+#pragma once
+
+#include <wx/dialog.h>
+
+class CGameListCtrl;
+class wxListBox;
+
+class ChangeGameDialog final : public wxDialog
+{
+public:
+	ChangeGameDialog(wxWindow* parent, const CGameListCtrl* const game_list, wxString& game_name);
+
+private:
+	void OnPick(wxCommandEvent& event);
+
+	wxListBox* m_game_lbox;
+	wxString&  m_game_name;
+};
